backend2/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlmodel==0.0.14
sqlalchemy==2.0.23
pydantic==2.5.0
python-dotenv==1.0.0
python-multipart==0.0.6
PyJWT==2.8.0
python-jose==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==25.1.0
pytest==7.4.3
httpx==0.25.2
psycopg2-binary==2.9.9
